Updated update, and getProductByCategoryId of CategoryController
