+++ api/src/main/java/com/community/api/entity/CustomProduct.java
@@ -47,33 +47,4 @@ public class CustomProduct extends ProductImpl {
         this.priorityLevel = priorityLevel;
     }
 
-    public static class CustomCategoryDao {
-        String categoryName;
-        Long categoryId;
-        List<Product> products;
-
-        public String getCategoryName() {
-            return this.categoryName;
-        }
-
-        public void setCategoryName(String categoryName) {
-            this.categoryName = categoryName;
-        }
-
-        public Long getCategoryId() {
-            return this.categoryId;
-        }
-
-        public void setCategoryId(Long categoryId) {
-            this.categoryId = categoryId;
-        }
-
-        public List<Product> getProducts() {
-            return this.products;
-        }
-
-      void setProducts(List<Product> products) {
-            this.products = products;
-        }
-    }
 }
