+++ api/src/main/java/com/community/api/services/CategoryService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -13,13 +14,13 @@ public class CategoryService {
     @PersistenceContext
     private EntityManager entityManager;
 
-    public List<CategoryProductXref> getAllProductsByCategoryId(Long categoryId) {
-        String sql = "SELECT `product_id` FROM blc_category_product_xref WHERE category_id = :categoryId";
+    public List<BigInteger> getAllProductsByCategoryId(Long categoryId) {
+        String sql = "SELECT product_id FROM blc_category_product_xref WHERE category_id = :categoryId";
 
         try {
             return entityManager.createNativeQuery(sql).setParameter("categoryId", categoryId).getResultList();
         } catch (Exception e) {
-            throw new RuntimeException("Failed to Delete Category_Product: " + e.getMessage(), e);
+            throw new RuntimeException("Failed to get Category_Product: " + e.getMessage(), e);
         }
     }
 }
